Terraform build and clean
